+++ test/web/core.js
@@ -17,8 +17,8 @@ success
     "content-type": "application/json",
     link: [
       {href: "/", rel: "self current"},
-      {href: "/foo", rel: "collection", title: "foo"},
-      {href: "/{title}", rel: "collection"}
+      {href: "/foo", id: "foo", rel: "collection"},
+      {href: "/{id}", rel: "collection"}
     ]
   },
   reason: "Ok",
@@ -62,9 +62,9 @@ success
     "content-type": "text/plain",
     link: [
       {href: "/", rel: "self current"},
-      {href: "/events", rel: "collection", title: "events"},
-      {href: "/foo", rel: "collection", title: "foo"},
-      {href: "/{title}", rel: "collection"}
+      {href: "/events", id: "events", rel: "collection"},
+      {href: "/foo", id: "foo", rel: "collection"},
+      {href: "/{id}", rel: "collection"}
     ]
   },
   reason: "ok",