+++ CHANGES.md
@@ -2,6 +2,11 @@ Changes
 =======
 0.4.0
 
+2013/06/19 pfraze
+
+ - Changed web.navigator (discovery protocol) to use the id attribute instead of title
+
+
 2013/06/03 pfraze
 
  - Added support for SharedWorkers